rename add posts handler to show posts
