+++ server.js
@@ -1,11 +1,11 @@
 const express = require("express");
 const home = require("./routes/home.js");
-const addPost = require("./routes/addPost.js");
+const showPosts = require("./routes/showPosts.js");
 
 const server = express();
 
 server.get("/", home.get);
-server.get("/add-post", addPost.get);
+server.get("/show-posts", showPosts.get);
 
 const PORT = process.env.PORT || 3000;
